requirements.txt
Django==5.2.6
supabase==2.18.1
python-dotenv==1.1.1
